

\name{krons}
\alias{krons}
\title{Calculates Kronecker Products using Sparse Matrices}

\description{Given a vector of coefficients and a list of B-spline basis functions, computes approximated function value. Uses sparsity of Basis functions.}

\usage{tbd}

\arguments{tbd}

\details{tbd}

\value{}
\author{<florian.oswald@gmail.com>}

\examples{
tbd}
	

